Add notification microservice

+++ clients/src/main/java/com/fvucemilo/clients/fraud/FraudClient.java
@@ -12,5 +12,5 @@ public interface FraudClient {
 
     @GetMapping(path = "api/v1/fraud-check/{customerId}")
     FraudCheckResponse isFraudster(
-            @PathVariable("customerId") Integer customerId);
+            @PathVariable("customerId") Long customerId);
 }
